all commented

+++ back/controllers/userController.js
@@ -1,8 +1,13 @@
+/*
+* Controller for users.
+*/
+
 'use strict';
 
 const {getAllUsers, getUser, insertUser, deleteUser, updateUser, updateToken} = require("../models/userModel");
 const {makeThumbnail} = require("../utils/resize");
-const {removeFile, removeFiles} = require("../utils/removeFile");
+const {removeFile} = require("../utils/removeFile");
+
 const user_list_get = async (req, res) => {
     const users = await getAllUsers();
     res.json(users);
@@ -13,6 +18,7 @@ const user_get = async (req, res) => {
     res.json(user);
 };
 
+// creating a thumbnail for the profile pic if one was uploaded
 const user_post = async (req, res) => {
     console.log('add user data ', req.body);
     console.log('profile pic ', req.file);
@@ -28,12 +34,15 @@ const user_post = async (req, res) => {
     }
 }
 
+// on delete, the profile pic and its thumbnail will be deleted from the server.
 const user_delete = async (req , res) => {
     req.body.userId = req.user.userId;
     try {
         const user = await getUser(req.user.userId);
-        const removed = await removeFile('./uploads/user/', './thumbnails/user/', user.profilePic);
-        console.log('profilePic removed: ' + removed);
+        if (user.profilePic) {
+            const removed = await removeFile('./uploads/user/', './thumbnails/user/', user.profilePic);
+            console.log('profilePic removed: ' + removed);
+        }
     } catch (e) {
         console.error(e)
     }
@@ -41,14 +50,18 @@ const user_delete = async (req , res) => {
     res.json(id);
 }
 
+// If a profile pic is added, the previous one will be removed from the server with its thumbnail.
+// A thumbnail for the new picture will be created.
 const user_update = async (req, res) => {
     console.log('req.body', req.body);
     req.body.userId = req.user.userId;
     if (req.file) {
         const currentUserInfo = await getUser(req.user.userId);
         try {
-            const removed = await removeFile('./uploads/user/', './thumbnails/user/', currentUserInfo.profilePic);
-            console.log('file removed: ', removed);
+            if (currentUserInfo.profilePic) {
+                const removed = await removeFile('./uploads/user/', './thumbnails/user/', currentUserInfo.profilePic);
+                console.log('file removed: ', removed);
+            }
         } catch (e) {
             console.error(e)
         }
@@ -73,6 +86,7 @@ const checkToken = (req, res, next) => {
     }
 };
 
+// returns user credentials to login again and update token
 const refreshToken = async (req, res) => {
     const user = await updateToken(req.user)
     res.json(user);
